Prevent auto changing content when facet selected
