+++ themes/stanford_basic/algolia-search/src/algolia-search.island.tsx
@@ -1,12 +1,7 @@
 import algoliasearch from 'algoliasearch/lite';
 import {createIslandWebComponent} from 'preact-island'
-import {
-  CurrentRefinements,
-  HitsProps,
-  InstantSearch, useCurrentRefinements,
-  useHits, useRefinementList
-} from 'react-instantsearch';
-import SearchBox from "./search-box";
+import {HitsProps, InstantSearch, useHits} from 'react-instantsearch';
+import SearchForm from "./search-form";
 import EventHit from "./hits/events";
 import NewsHit from "./hits/news";
 import DefaultHit from "./hits/default-hit";
@@ -55,47 +50,6 @@ const CustomHits = (props) => {
   )
 }
 
-const CustomRefinementList = (props) => {
-  const {items, refine, canRefine, createURL} = useRefinementList(props);
-
-  const onChange = (item) => {
-    refine(item.value)
-  }
-  return (
-    <ul style={{listStyle: "none"}}>
-      {items.sort((a, b) => a.count < b.count ? 1: (a.count === b.count ? (a.value < b.value ? -1 : 1): -1)).map((item, i) =>
-        <li key={i}>
-          <label>
-            <input
-              type="checkbox"
-              disabled={!canRefine}
-              checked={item.isRefined}
-              onChange={() => onChange(item)}
-            />
-            {item.value} ({item.count})
-          </label>
-        </li>
-      )}
-    </ul>
-  )
-}
-
-const CustomCurrentRefinements = (props)=> {
-  const { items, canRefine, refine } = useCurrentRefinements(props);
-
-  return (
-    <ul style={{listStyle:"none", display: "flex", flexWrap: "wrap", gap:"10px"}}>
-      {items.map(refinement => {
-       return refinement.refinements.map((item, i) =>
-          <li key={`refinement-${i}`}>
-            {item.value}
-            <button disabled={!canRefine} onClick={() => refine(item)}>Clear</button>
-          </li>
-        )
-      })}
-    </ul>
-  );
-}
 
 const Search = () => {
   const currentUrl = new URL(window.location.href);
@@ -111,9 +65,7 @@ const Search = () => {
       }}
     >
       <Container>
-        <SearchBox/>
-        <CustomRefinementList attribute="news_type"/>
-        <CustomCurrentRefinements/>
+        <SearchForm/>
         <CustomHits/>
       </Container>
     </InstantSearch>
